refactor(286): rework after review
