+++ src/features/student/queries/index.ts
@@ -1,4 +1,4 @@
+export * from './fixtures'
 export * from './use-program-progress-view.query/use-program-progress-view.query'
 export * from './use-student-access-control.query/use-student-access-control.query'
 export * from './use-student-summary.query/use-student-summary.query'
-export * from './utils'
