+++ src/features/student/queries/use-program-progress-view.query/use-program-progress-view.query.ts
@@ -1,6 +1,6 @@
 import type { BaseApiException } from '@/common/exceptions'
 import type { ProgramProgressViewDTO } from '@/types'
-import { mockedPrograms } from '@/features/student/queries/utils'
+import { mockedPrograms } from '@/features/student/queries/fixtures'
 import { useQuery, type UseQueryReturnType } from '@tanstack/vue-query'
 
 const commonQueryKeys = ['user', 'student']
