+++ src/features/student/queries/use-student-summary.query/use-student-summary.query.test.ts
@@ -6,7 +6,7 @@ import type { BaseApiException } from '@/common/exceptions'
 import type { UseQueryReturnType } from '@tanstack/vue-query'
 import profile_banner_placeholder from '@/assets/profile_banner_placeholder.png'
 import profile_picture_placeholder from '@/assets/profile_picture_placeholder.png'
-import { mockedCourses } from '@/features/student/queries/utils'
+import { mockedCourses } from '@/features/student/queries/fixtures'
 import {
   type DeliverableOverviewDTO,
   type EventOverviewDTO,
