+++ src/ui/styles/palette.scss
@@ -4,7 +4,7 @@
   --background-dialog: #FDFDFD;
   --background-surface-background: #F6F6F6;
 
-  --dark-background-critical: #9E0592; //#8B64F4; // it's the old one but not contrasted enough
+  --dark-background-critical: #8B64F4;
   --dark-background-primary1: #000091;
   --dark-background-warn: #9C3900;
 
@@ -24,6 +24,7 @@
   --light-background-warn: #F4CEB8;
 
   --light-foreground-error: #800108;
+  --light-foreground-primary1: #2929A2;
   --light-foreground-success: #0F4625;
 
   --skill1: #7216D4;
